requirements.txt
keras==2.9.0
matplotlib==3.5.0
numpy==1.21.4
Pillow==9.1.1
tensorflow==2.9.0
